Refactoring: Introduce signaling trait (#2)
This is a big refactoring that converts the current enum based signaling implementation to one where we have a Signaling trait.

Common functionality is moved from the `Signaling` enum both into the trait as well as into a `Common` struct.

The `TestContext` now has a type parameter. This gets rid of some ugly unwrapping.

Overall, the codebase is now much nicer! ✨