Back to async (#31)
The move to a synchronous API (#29) has shown a few problems:

- Websocket library doesn't seem to be actively maintained anymore
- Websocket library uses openssl crate instead of native-tls
- Websocket library does not report connection errors to user code
- More threads required than expected

Because the previous approach at an async API for saltyrtc-client
failed, I used a new approach with forwarding a stream to a sink,
without using `loop_fn`. External messages are selected from a futures
channel. This approach is cleaner and supports our use case better,
without passing around a lot of state!